listener: implement an abstract listener
Implement an abstract SensateIoT websocket listener.
