+++ SensateIoT.SmartEnergy.Dsmr.WebClient.Common/SensateIoT.SmartEnergy.Dsmr.WebClient.Common.csproj
@@ -56,7 +56,7 @@
   <ItemGroup>
     <Compile Include="Abstract\IDsmrParserService.cs" />
     <Compile Include="Abstract\IMeasurementStorageService.cs" />
-    <Compile Include="Abstract\IWebSocketClientService.cs" />
+    <Compile Include="Abstract\IListener.cs" />
     <Compile Include="Connected Services\WCF.Parser\Reference.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -66,11 +66,12 @@
     <Compile Include="Events\EventType.cs" />
     <Compile Include="Events\WebSocketEventArgs.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\BaseListener.cs" />
     <Compile Include="Services\DsmrParserService.cs" />
     <Compile Include="Services\MeasurementStorageService.cs" />
     <Compile Include="Services\PingService.cs" />
     <Compile Include="Services\AuthorizationService.cs" />
-    <Compile Include="Services\WebSocketClientService.cs" />
+    <Compile Include="Services\WebSocketListenerService.cs" />
     <Compile Include="Settings\ParserSettings.cs" />
     <Compile Include="Telegram.cs" />
   </ItemGroup>
