+++ SensateIoT.SmartEnergy.Dsmr.WebClient.Service/SensateIoT.SmartEnergy.Dsmr.WebClient.Service.csproj
@@ -76,6 +76,7 @@
     <Compile Include="Application\ConsoleHost.cs" />
     <Compile Include="Application\ConfigurationLoader.cs" />
     <Compile Include="Application\Program.cs" />
+    <Compile Include="Clients\DsmrClient.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\WebClientService.cs" />
     <Compile Include="Services\WindowsService.cs">
